+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/entity/statistics/CriterionProductStatisticsEntity.java
@@ -1,0 +1,77 @@
+package gov.healthit.chpl.entity.statistics;
+
+import gov.healthit.chpl.certificationCriteria.CertificationCriterionEntity;
+import gov.healthit.chpl.domain.CriterionProductStatistics;
+import gov.healthit.chpl.entity.EntityAudit;
+import gov.healthit.chpl.entity.lastmodifieduserstrategy.CurrentUserThenSystemUserStrategy;
+import gov.healthit.chpl.entity.lastmodifieduserstrategy.LastModifiedUserStrategy;
+import jakarta.persistence.Basic;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@Setter
+@ToString
+@SuperBuilder
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "criterion_product_statistics")
+public class CriterionProductStatisticsEntity extends EntityAudit {
+    private static final long serialVersionUID = -4258273713908999510L;
+
+    @Override
+    public LastModifiedUserStrategy getLastModifiedUserStrategy() {
+        return new CurrentUserThenSystemUserStrategy();
+    }
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Basic(optional = false)
+    @Column(name = "id", nullable = false)
+    private Long id;
+
+    @Basic(optional = false)
+    @Column(name = "product_count", nullable = false)
+    private Long productCount;
+
+    @Basic(optional = false)
+    @Column(name = "certification_criterion_id", nullable = false)
+    private Long certificationCriterionId;
+
+    @OneToOne(optional = true, fetch = FetchType.LAZY)
+    @JoinColumn(name = "certification_criterion_id", insertable = false, updatable = false)
+    private CertificationCriterionEntity certificationCriterion;
+
+    public CriterionProductStatisticsEntity(final Long id) {
+        this.id = id;
+    }
+
+    @Transient
+    public Class<?> getClassType() {
+        return CriterionProductStatisticsEntity.class;
+    }
+
+    public CriterionProductStatistics toDomain() {
+        return CriterionProductStatistics.builder()
+                .id(this.id)
+                .productCount(this.productCount)
+                .certificationCriterionId(this.certificationCriterionId)
+                .criterion(this.certificationCriterion.toDomain())
+                .build();
+    }
+}