+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/search/entity/ListingSearchEntity.java
@@ -166,4 +166,7 @@ public class ListingSearchEntity {
 
     @Column(name = "children")
     private String children;
+
+    @Column(name = "standards_met")
+    private String standardsMet;
 }